♻ changed from [...slug] to [proj]/[year] & used ssg
