+++ src/app/(main)/projects/[proj]/_components/IndividualProject.tsx
@@ -74,4 +74,4 @@ const IndividualProject = ({ project }: IndividualProjectProps) => {
     );
 };
 
-export default IndividualProject;
+export default IndividualProject;
